Revert most of 9b96703cd3b1b33bb706ed49df71c76d7978ae4e
Decided to put the rule in the title and the url where it normally is. jsUri was giving me grief so I just did it the easy way and passed a encoded JSON object. Everything is good now, except those hours and hours I spent making the origin solution work ;_;